feat: Update WorkLog, Add Check MaxTime and MinTime

+++ BE/CapstoneProject_SP25_IPAS/CapstoneProject_SP25_IPAS_Service/Service/WorkLogService.cs
@@ -109,6 +109,7 @@ namespace CapstoneProject_SP25_IPAS_Service.Service
                     StartDate = addNewTaskModel.DateWork.Value.Add(TimeSpan.Parse(addNewTaskModel.StartTime)),
                     EndDate = addNewTaskModel.DateWork.Value.Add(TimeSpan.Parse(addNewTaskModel.EndTime)),
                     Frequency = "None",
+                    MasterTypeId = getMasterType.MasterTypeId,
                     IsActive = true,
                     IsDeleted = false,
                     Status = "Active",
@@ -639,6 +640,26 @@ namespace CapstoneProject_SP25_IPAS_Service.Service
                             throw new Exception("Start time must be less than End Time");
                         }
                     }
+                    if (updateWorkLogModel.StartTime != null && updateWorkLogModel.EndTime != null)
+                    {
+                        var parseStartTime = TimeSpan.Parse(updateWorkLogModel.StartTime);
+                        var parseEndTime = TimeSpan.Parse(updateWorkLogModel.EndTime);
+                        var checkTime = (int)(parseEndTime - parseStartTime).TotalHours; // Chuyển TimeSpan sang số phút
+
+                        var masterType = await _unitOfWork.MasterTypeRepository
+                            .GetByCondition(x => x.MasterTypeId == updateWorkLogModel.MasterTypeId);
+
+                        if (masterType != null)
+                        {
+                            var minTime = masterType.MinTime;
+                            var maxTime = masterType.MaxTime;
+
+                            if (checkTime < minTime || checkTime > maxTime)
+                            {
+                                throw new Exception($"Time of work ({checkTime} hours) does not valid! It must be in range {minTime} - {maxTime} hours.");
+                            }
+                        }
+                    }
                     var checkExistProcess = await _unitOfWork.ProcessRepository.GetByCondition(x => x.ProcessId == updateWorkLogModel.ProcessId);
                     if (updateWorkLogModel.DateWork != null)
                     {
@@ -1068,6 +1089,24 @@ namespace CapstoneProject_SP25_IPAS_Service.Service
                 {
                     return new BusinessResult(400, "Plan does not exist");
                 }
+                if (addNewTaskModel.StartTime != null && addNewTaskModel.EndTime != null)
+                {
+                    var checkTime = (int)(endTime - startTime).TotalHours; // Chuyển TimeSpan sang số phút
+
+                    var masterType = await _unitOfWork.MasterTypeRepository
+                        .GetByCondition(x => x.MasterTypeId == getExistPlan.MasterTypeId);
+
+                    if (masterType != null)
+                    {
+                        var minTime = masterType.MinTime;
+                        var maxTime = masterType.MaxTime;
+
+                        if (checkTime < minTime || checkTime > maxTime)
+                        {
+                            throw new Exception($"Time of work ({checkTime} hours) does not valid! It must be in range {minTime} - {maxTime} hours.");
+                        }
+                    }
+                }
 
                 await _unitOfWork.WorkLogRepository.CheckWorkLogAvailabilityWhenAddPlan(
                                                                       startTime,
@@ -1477,10 +1516,33 @@ namespace CapstoneProject_SP25_IPAS_Service.Service
                     {
                         if (changeEmployee.Status.ToLower().Equals("add"))
                         {
-                            getUserToUpdate.StatusOfUserWorkLog = WorkLogStatusConst.REJECTED;
-                            getUserToUpdate.ReplaceUserId = changeEmployee.NewUserId;
+                            // Nếu đã có người thay thế trước đó, cập nhật người mới thay vì insert liên tục
+                            if (getUserToUpdate.ReplaceUserId != null)
+                            {
+                                // Xóa user thay thế trước đó để tránh dư thừa dữ liệu
+                                var existingReplacement = await _unitOfWork.UserWorkLogRepository
+                                    .GetByCondition(x => x.UserId == getUserToUpdate.ReplaceUserId
+                                                         && x.WorkLogId == getUserToUpdate.WorkLogId);
+
+                                if (existingReplacement != null)
+                                {
+                                    _unitOfWork.UserWorkLogRepository.Delete(existingReplacement);
+                                    await _unitOfWork.SaveAsync();
+                                }
+
+                                // Cập nhật lại ReplaceUserId thành người mới
+                                getUserToUpdate.ReplaceUserId = changeEmployee.NewUserId;
+                            }
+                            else
+                            {
+                                getUserToUpdate.StatusOfUserWorkLog = WorkLogStatusConst.REJECTED;
+                                getUserToUpdate.ReplaceUserId = changeEmployee.NewUserId;
+                            }
+
                             _unitOfWork.UserWorkLogRepository.Update(getUserToUpdate);
                             await _unitOfWork.SaveAsync();
+
+                            // Chỉ tạo mới người cuối cùng
                             var newUserWorkLog = new UserWorkLog()
                             {
                                 CreateDate = DateTime.Now,
@@ -1490,14 +1552,22 @@ namespace CapstoneProject_SP25_IPAS_Service.Service
                                 IsDeleted = false,
                                 StatusOfUserWorkLog = WorkLogStatusConst.REPLACED,
                             };
-                            await _unitOfWork.UserWorkLogRepository.Insert(newUserWorkLog);
 
+                            await _unitOfWork.UserWorkLogRepository.Insert(newUserWorkLog);
                         }
+
                         else if (changeEmployee.Status.ToLower().Equals("update"))
                         {
+                            // Nếu đổi rồi nhưng cuối cùng chọn lại người cũ thì không cần thay đổi
+                            if (changeEmployee.NewUserId == getUserToUpdate.UserId)
+                            {
+                                continue; // Không làm gì cả
+                            }
 
+                            // Nếu thực sự đổi người mới
                             _unitOfWork.UserWorkLogRepository.Delete(getUserToUpdate);
                             await _unitOfWork.SaveAsync();
+
                             var newUserWorkLog = new UserWorkLog()
                             {
                                 CreateDate = DateTime.Now,
@@ -1509,6 +1579,7 @@ namespace CapstoneProject_SP25_IPAS_Service.Service
                             };
                             await _unitOfWork.UserWorkLogRepository.Insert(newUserWorkLog);
                         }
+
                     }
                     else
                     {
